Share core test execution logic between native and WASM paths
Native and WASM/Pyodide were diverging into two parallel implementations
of the same plumbing. Push every duplicated piece down to its natural
common location so each surface keeps only the thin glue it actually
needs.

- `tryke_types` now owns `convert_wire_result`, the full assertion
  enrichment pipeline (`convert_assertion`, span computation, expected
  matching, executed-line mapping, etc.) and a `RunSummary::from_results`
  helper. Both `tryke_runner` and `tryke_wasm` call into them instead of
  carrying private copies.
- `tryke_wasm` drops its hand-rolled `convert_wire_result_simple` and
  `build_summary`, and stops needing `TestOutcome` / `Duration` imports.
- `python/tryke/runner.py` grows `HookInfo`, `build_executor_from_hooks`,
  and `run_doctest` — all derived directly from the worker/playground
  duplicates. `worker.py` keeps its JSON-RPC validation step and then
  delegates; `playground.py` becomes a thin harness around
  `run_doctest` / `build_executor_from_hooks` / `run_test`.
- Drop the now-unused `span_offset` / `span_length` fields from
  `AssertionWire`: they were always `0` from Python and overwritten by
  the Rust-side `compute_subject_span`, so the wire roundtrip was
  carrying dead weight.
- `playground.py` reads `case_label` straight off the `TestItem`
  (discovery already emits it) instead of looking it up through
  `__tryke_cases__` by `case_index`.
